Adds module utils.js and data.js
